+++ services/src/main/java/org/fao/geonet/api/registries/DirectoryApi.java
@@ -64,7 +64,6 @@ import org.geotools.feature.FeatureIterator;
 import org.geotools.feature.SchemaException;
 import org.geotools.feature.simple.SimpleFeatureBuilder;
 import org.geotools.geometry.jts.JTS;
-import org.geotools.gml.producer.GeometryTransformer;
 import org.geotools.referencing.CRS;
 import org.jdom.Element;
 import org.opengis.feature.simple.SimpleFeature;
@@ -98,7 +97,6 @@ import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FilenameFilter;
 import java.io.IOException;
-import java.net.MalformedURLException;
 import java.net.URISyntaxException;
 import java.net.URL;
 import java.nio.charset.Charset;
