+++ services/src/test/resources/org/fao/geonet/api/registries/layers.zip
